requirements.txt
fastapi
uvicorn
python-multipart
pillow
langchain
openai
transformers
torch